+++ internal/k8s/export_test.go
@@ -2,4 +2,6 @@ package k8s
 
 var ParsePodEnvironment = parsePodEnvironment
 
+var GetKubectlVersion = getKubectlVersion
+
 var NewExecutor = newExecutor